fix(swipeable-routes): dynamically set height of swipeable route container based on content/screen
the height of the swipeable container would be the height of the route with the most content. This
allowed for scrolling on routes even when there was no content to show. In this commit the height of
hidden routes is set to 0 while the height of the visible route is set to 100%. This means the
height of the container will always equal the height of the content being shown or the height of the screen.

Refs #65