refactor(location): move LocationAutocomplete to a REST endpoint
Location can now also be sent through Create Event to the backend. Also added ToastOptions to useToast.

With the GraphQL endpoint, in order to use IGooglePlace as a return type for the autocompleteLocations request, we needed to create a new class called GooglePlace annotated with the decorator @ObjectType. This ended up being convoluted
for no reason, therefore we ended up moving everything to a REST endpoint.

Refs #124

Co-authored-by: Cristian Aldea <aldea.cristian@outlook.com>