feat(event): Fix event page host placeholder
Fixed the event page to show the name of the user that hosts the event

To do this, all queries in the event module now also return host information. This was done through adding host as a relation, or adding a left join when the query builder is used.

Also, took the liberty of changing the GraphQL type of ids to Int, which avoids us having to parse it when fetching

ref #171