fix(ml): Fix wrong order of model output
Initially, we assumed that the order of the tensors within the model output were the same. however, after uploading a new model everything broke. The order of the tensors aren't determined. this was fixed by explicitly asking the model for the tensors we need by name (the names are obfuscated/lost, so a comment clarifies what each tensor is).