fix(hostname): Allow easy hostname change through host.env.ts
The host.env.ts file was added to allow the developer to easily change the hostname where their application is being served, to allow for usecases where they don't want to run on localhost. For example, if they want other devices on their local network to be able to acce
ss the application, using localhost won't work, and they would need to host in on their machine's ip address for example

The README was updated to include info on how this mechanism works

The reason I went with a .ts file instead of adding these variables to our .env file is that the frontend cannot read the environment variables of our webserver, since it is served to the client's browser. My solution therefore involved gitignoring a ts with the values, which would allow us to set these values without hardcoding them.

Ref #217