feat(ml): Implement click feature for ml page
A click listener was setup for the canvas in the waste recognition page, which triggers the onCanvasClick method. The click coordinates are the transformed to image space and the results founds since the last successful detection are looped through to find objects boxes which contain the click. The first result is considered, since results are ordered by confidence.

ref #230