feat(user resolver): Set removeFriend to be two-way
I realized something odd in my design of removeFriend. removeFriend won't remove both sides of the relationship. This makes sense when the user wants to, say, cancel a friend request, but it makes less sense when the two users share a mutual friendship. Therefore I decided to have the method just remove the friendship on both sides in all cases